Refactor CSS classes for consistency, add responsive design styles, and replace product images with SVGs
