+++ src/components/About/About.jsx
@@ -1,20 +1,20 @@
 import React from "react";
 import "./About.css";
 
-import main_photo1 from "../../assets/About-us-image-1.jpg";
-import main_photo2 from "../../assets/About-us-image-2.jpg";
-import main_photo3 from "../../assets/About-us-image-3.jpg";
+import main_photo1 from "../../assets/About-us-image-1.svg";
+import main_photo2 from "../../assets/About-us-image-2.svg";
+import main_photo3 from "../../assets/About-us-image-3.svg";
 function About() {
   return (
-    <div className="main__page">
+    <div className="about__page">
       <h1>About Us</h1>
 
-      <div className="main__photos_container">
-        <img src={main_photo1} className="main__photos" alt="first photo" />
-        <img src={main_photo2} className="main__photos"  alt="second photo" />
-        <img src={main_photo3} className="main__photos"  alt="third photo" />
+      <div className="about__photos-container">
+        <img src={main_photo1} className="about__photos" alt="first photo" />
+        <img src={main_photo2} className="about__photos"  alt="second photo" />
+        <img src={main_photo3} className="about__photos"  alt="third photo" />
       </div>
-      <p className="main__text">
+      <p className="about__text">
         At Imena Jewelry Manufacturing Inc., our passion for crafting exquisite,
         high-end jewelry is deeply rooted in our family’s legacy of artistry and
         dedication. For generations, we have embraced the art of jewelry making,
@@ -29,7 +29,7 @@ function About() {
         define Imena Jewelry Manufacturing Inc.
       </p>
       {/* <img src={main_photo1} className="main__media" alt="main photo 1" /> */}
-      <div className="main__page-text"></div>
+      <div className="about__page-text"></div>
     </div>
   );
 }
