+++ src/components/App/App.css
@@ -1,5 +1,14 @@
+html, body {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  display:flex;
+  justify-content: center;
+  
+}
+
 .page {
-  width: 1440px;
+  width: 100%;
   height: 100%;
   background-color: #fff;
 }
